+++ openapi_python_client/openapi_parser/errors.py
@@ -1,0 +1,10 @@
+from typing import Any, Optional
+
+
+class ParseError(ValueError):
+    """ An error raised when there's a problem parsing an OpenAPI document """
+
+    def __init__(self, data: Any, message: Optional[str] = None):
+        super().__init__()
+        self.data = data
+        self.message = message